+++ pkg/scalers/scalersconfig/typed_config_test.go
@@ -610,15 +610,56 @@ func TestMultiName(t *testing.T) {
 	Expect(ts.Property).To(Equal("bbb"))
 }
 
+// TestUnexpectedOptional tests the unexpected optional input
+func TestUnexpectedOptional(t *testing.T) {
+	RegisterTestingT(t)
+
+	// Create a mock recorder to capture the event
+	mockRecorder := &MockEventRecorder{Messages: make([]string, 0)}
+
+	sc := &ScalerConfig{
+		TriggerMetadata: map[string]string{
+			"stringVal":     "value1",
+			"nestVal":       "nestVal",
+			"notexistVal":   "aaa",
+			"notValFromEnv": "bbb",
+		},
+		ResolvedEnv: map[string]string{
+			"bbb": "1.1",
+		},
+		Recorder:           mockRecorder,
+		ScalableObjectType: "test",
+	}
+	type nestStruct struct {
+		Nest string `keda:"name=nestVal, order=triggerMetadata"`
+	}
+
+	type testStruct struct {
+		BA        nestStruct `keda:""`
+		StringVal string     `keda:"name=stringVal, order=triggerMetadata"`
+	}
+
+	ts := testStruct{}
+	err := sc.TypedConfig(&ts)
+	Expect(err).To(BeNil())
+
+	// Verify that the event was recorded
+	Expect(mockRecorder.EventCalled).To(BeTrue())
+	Expect(mockRecorder.Messages[0]).To(Equal("Unmatched input property notexistVal in scaler test"))
+	Expect(mockRecorder.Messages[1]).To(Equal("Unmatched input property notValFromEnv in scaler test"))
+}
+
 // MockEventRecorder is a mock implementation of record.EventRecorder
 type MockEventRecorder struct {
 	EventCalled bool
 	Message     string
+	Messages    []string
 }
 
 func (m *MockEventRecorder) Event(object runtime.Object, eventtype, reason, message string) {
 	m.EventCalled = true
 	m.Message = message
+	m.Messages = append(m.Messages, message)
 }
 
 func (m *MockEventRecorder) Eventf(object runtime.Object, eventtype, reason, messageFmt string, args ...interface{}) {